Conf and snapshot paths can now be different
Changed how remote paths are handled. OVZ-Backup no longer uses
hard-coded folders 'remote_path/snapshot' and 'remote_path/conf' for
storing backups. These folders had to be created manually or rsync
would fail. The full path must now be given as a command line
argument. Up to two paths can be provided. The first path argument is
used for storing snapshots, and the second for configuration files. It
is also possible to only give it one path argument. If only one path
argument is used then it will be used for both the snapshots and
configuration files.

Put in more detailed comments to explain which path variable holds
which of the remote/local paths to configuration files and snapshots.